Add the possibility to specify parameters specific per hub/port
Add the possibility to specify via command line a parameter specific for
a given hub-port combination in the format
    -u parameterName[hub index][port index]

(whereas the already present option -s apply the same extra parameters
to all devices).

For example if we want to flash 4 mac address, each one different for any
device we can specify something like this:

 mfgtool2.exe -p 4 -c "linux" -l ""eMMC"" -s "mmc=0" -u
 mac[1][1]="AB:CD:EF:12:34:56" -u mac[1][2]="12:23:34:AA:BB:CC" -u
 mac[2][1]="AA:BB:CC:DD:EE:FF" -u mac[3][1]= "11:22:33:44:55:66"

Then in ucl2.xml we can just use %parameter% (%mac% in our example
above) to reference a specific parameter, and depending on the hub -
port indexes, the proper substition will be taken in place e.g. if the
device is inserted in hub =2, port =1 we will replace %mac% with
"AA:BB:CC:DD:EE:FF".

Signed-off-by: Paolo La Camera <paolo.lacamera@gmail.com>